Eliminating the need for files and actions in yaml
Update to eliminate the need for explicitly declaring the files and actions sections in the build-info.yaml file.

The script will automatically detect the contents of the payload folder and any available preinstall and postinstall scripts in the scripts folder.

This way, the script can dynamically generate the required WiX configuration without the user needing to manually define these sections in the YAML file.