`CustomAction` changes when there's no script
Invalid Directory ID and Name: The Id and Name attributes for the directory C:\ are not legal in WiX as they contain characters that aren't allowed in identifiers.

Deprecated VBScript: The use of VBScript as a placeholder for CustomAction is causing issues, as VBScript is deprecated and may not work on all systems.

StandardDirectory Warning: The warning about TARGETDIR indicates that the StandardDirectory element should be used instead.